Created new confirmation dialog; cleared some redundancy in dialog usage
